Add development mode and fix dev status task
- Put large gulp task into their own file
- Add 'npm run dev'
- Prevent changes to comparison.yml
- Auto configuration tasks now change comparison-auto-config.yml
- Change dev status task from sync to async
- Handle errors in git api responses
