+++ lib/gulp/gulpfile.js
@@ -1,10 +1,3 @@
-require("@babel/register")({
-    only: [
-        /node_modules\/ultimate-comparison\/lib\/gulp\/gulpfile.babel.js/,
-        /lib\/gulp\/gulpfile.babel.js/,
-        /lib\\gulp\\gulpfile.babel.js/,
-        /node_modules\\ultimate-comparison\\lib\\gulp\\gulpfile.babel.js/
-    ]
-});
+require("@babel/register");
 
 module.export = require("./gulpfile.babel.js");
