+++ lib/gulp/tasks/citation.babel.js
@@ -1,0 +1,89 @@
+import {readFile, readFileSync, writeFileSync} from "fs";
+import yaml2json from "js-yaml";
+import * as path from "path";
+
+const Cite = require('citation-js');
+
+export function citation(paths, files, done) {
+    let input = yaml2json.safeLoad(readFileSync(files.autoConfig, "utf8"));
+    let defaults = yaml2json.safeLoad(readFileSync(files.defaultConfig, "utf8"));
+    const citation = input.citation || {};
+    const citationDefault = defaults.citation;
+    const csl = path.join(paths.data, (citation.csl || citationDefault.csl));
+    const bib = path.join(paths.data, (citation.bib || citationDefault.bib));
+
+    if (csl) {
+        readFile(csl, "utf8", function (err, cslString) {
+            if (err) {
+                return console.error("Could not read File: ".concat(err.toString()));
+            }
+            Cite.CSL.register.addTemplate("defaultParameter", cslString.toString());
+            readBib(done)
+        });
+    } else {
+        console.info("Undefined 'csl' file using default!");
+        readBib(done)
+    }
+
+    function readBib(done) {
+        if (bib) {
+            readFile(bib, "utf8", function (err, data) {
+                let changed;
+                if (err) {
+                    return console.error("Could not read File: ".concat(err.toString()));
+                }
+                if ((data || "").replace(/([^%]*)(%[^\n]*)([^%]*)/, "$1$3").trim().length === 0) {
+                    input.autoCitation = {};
+                    writeFileSync(files.autoConfig, yaml2json.safeDump(input), "utf8");
+                    return done();
+                }
+                const cite = new Cite(data.toString().replace(/^%.*\n?/gm, ''), {forceType: 'string/bibtex'});
+                let map = new Map();
+                for (let item of cite.data) {
+                    let itemData = new Cite(item);
+                    map.set(item.id, (csl ?
+                            itemData.get({
+                                type: 'string',
+                                style: 'citation-defaultParameter'
+                            }) :
+                            item.get({type: 'string'})
+                        ).trim().replace(/\. \./gm, ".")
+                    );
+                    changed = true;
+                }
+
+                if (changed) {
+                    let data = readFileSync(files.dataJson, "utf8");
+                    data = data.concat(readFileSync(files.description, "utf8"));
+                    let keys = new Set();
+                    let keyReg = /\[@(.*?)]/g;
+                    let match;
+                    do {
+                        match = keyReg.exec(data);
+                        if (match) keys.add(match[1]);
+                    } while (match);
+
+                    keys.forEach(key => {
+                        if (!map.has(key)) {
+                            throw new Error("Bibtex entry for key '".concat(key, "' is missing!"));
+                        }
+                    });
+
+                    let obj = Object.create(null);
+                    let i = 0;
+                    for (let [k, v] of map) {
+                        if (data.match('@'.concat(k))) {
+                            obj[k] = {index: i, value: v};
+                            i++;
+                        }
+                    }
+                    input.autoCitation = obj;
+                    writeFileSync(files.autoConfig, yaml2json.safeDump(input), "utf8");
+                }
+                done();
+            });
+        } else {
+            done();
+        }
+    }
+}
