+++ lib/gulp/tasks/criteria.babel.js
@@ -1,0 +1,101 @@
+import {readFileSync, writeFileSync} from "fs";
+import yaml2json from "js-yaml";
+import _ from 'lodash';
+
+export function criteria(files, done) {
+    let config = yaml2json.safeLoad(readFileSync(files.config, "utf8")) || {};
+    let autoConfig = yaml2json.safeLoad(readFileSync(files.autoConfig, "utf8")) || {};
+    const defaultConfig = yaml2json.safeLoad(readFileSync(files.defaultConfig, "utf8"));
+    const defaultCriteria = defaultConfig.autoCriteria || {};
+    let criteriaObject = config.criteria || [];
+    let criteria = new Map();
+    criteriaObject.forEach(criteriaMap => Object.keys(criteriaMap).forEach(key => {
+        const valueObject = criteriaMap[key] || {};
+        const valuesObject = valueObject.values || {};
+        let values = new Set();
+        Object.keys(valuesObject).forEach(valueKey => values.add(valueKey));
+        criteria.set(key, {type: valueObject.type || defaultConfig.criteria[0].Example.type, values: values});
+    }));
+
+    const data = JSON.parse(readFileSync(files.dataJson, "utf8")) || [];
+
+    let autoCriteria = Object.create(null);
+    data.forEach(entry => Object.keys(entry).forEach(entryKey => {
+        if ("tag" === entryKey.toString() || "descr" === entryKey.toString()) {
+            if (!criteria.has("id")) {
+                autoCriteria["id"] = defaultCriteria.id;
+            }
+
+            if (!criteria.has("description")) {
+                autoCriteria["description"] = defaultCriteria.description;
+            }
+            return;
+        }
+
+        const entryValue = entry[entryKey];
+        if (criteria.has(entryKey) && "label" !== criteria.get(entryKey).type) {
+            return;
+        } else if (criteria.has(entryKey)) {
+            const childs = entryValue.childs || {};
+            const firstChild = childs["0"] || [];
+            firstChild.forEach(array => {
+                const vals = array || [];
+                vals.forEach(value => {
+                    const name = value.content;
+                    if (!name) return;
+                    if (!criteria.get(entryKey).values.has(name)) {
+                        if (!autoCriteria[entryKey]) {
+                            autoCriteria[entryKey] = {};
+                            autoCriteria[entryKey].values = {};
+                        }
+                        autoCriteria[entryKey].values[name] = {};
+                    }
+                });
+
+            });
+        } else {
+            const childs = entryValue.childs || {};
+            const firstChild = childs["0"] || [];
+            if (firstChild.length > 0) {
+                if (firstChild[0] === "") {
+                    return;
+                }
+                if (typeof firstChild[0] === "string") {
+                    autoCriteria[entryKey] = _.cloneDeep(defaultCriteria.defaultMarkdown);
+                } else {
+                    if (autoCriteria[entryKey] === undefined) {
+                        autoCriteria[entryKey] = _.cloneDeep(defaultCriteria.defaultLabel);
+                        autoCriteria[entryKey].values = {};
+                    }
+                    firstChild.forEach(array => {
+                        const vals = array || [];
+                        vals.forEach(value => {
+                            const name = value.content;
+                            if (!name) return;
+                            autoCriteria[entryKey].values[name] = {};
+                        });
+                    })
+                }
+            }
+
+        }
+    }));
+
+    // Renew 'comparison-auto-config.yml' with with 'comparison.yml', autoCriteria, autoColor and autoCitation
+
+    // Change column types from repository to label
+    let autoCriteriaObject = config.criteria || [];
+    autoCriteriaObject.forEach(criteriaMap => Object.keys(criteriaMap).forEach(key => {
+        const valueObject = criteriaMap[key] || {};
+        if (valueObject.type === 'repository') {
+            valueObject.type = 'label';
+        }
+    }));
+
+    config.autoCriteria = autoCriteria;
+    config.autoColor = autoConfig.autoColor;
+    config.autoCitation = autoConfig.autoCitation;
+    writeFileSync(files.autoConfig, yaml2json.safeDump(config), "utf8");
+
+    done();
+}
