+++ lib/gulp/tasks/determineColors.babel.js
@@ -1,0 +1,178 @@
+import {readFileSync, writeFileSync} from "fs";
+import yaml2json from "js-yaml";
+import _ from 'lodash';
+
+export function determineColors(files, done) {
+    const config = files.autoConfig;
+    const colorArray = [
+        'hsl(15, 100%, 70%)',
+        'hsl(30, 100%, 70%)',
+        'hsl(45, 100%, 70%)',
+        'hsl(60, 100%, 70%)',
+        'hsl(75, 100%, 70%)',
+        'hsl(90, 100%, 70%)',
+        'hsl(105, 100%, 70%)',
+        'hsl(120, 100%, 70%)',
+        'hsl(135, 100%, 70%)',
+        'hsl(150, 100%, 70%)',
+        'hsl(165, 100%, 70%)',
+        'hsl(180, 100%, 70%)',
+        'hsl(195, 100%, 70%)',
+        'hsl(210, 100%, 70%)',
+        'hsl(225, 100%, 70%)',
+        'hsl(240, 100%, 70%)',
+        'hsl(255, 100%, 70%)',
+        'hsl(270, 100%, 70%)',
+        'hsl(285, 100%, 70%)',
+        'hsl(300, 100%, 70%)',
+        'hsl(315, 100%, 70%)',
+        'hsl(330, 100%, 70%)'
+    ];
+    const foregroundArray = [
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#ffff00",
+        "#ffff00",
+        "#ffff00",
+        "#ffff00",
+        "#ffff00",
+        "#0d0d0d",
+        "#0d0d0d",
+        "#0d0d0d"
+    ];
+    let color;
+    let input = yaml2json.safeLoad(readFileSync(config, "utf8"));
+
+    const data = _.cloneDeep(input).criteria || [];
+    const autoCriteria = _.cloneDeep(input.autoCriteria) || {};
+    let changed = false;
+    let criteriaSet = new Set();
+    let criteriaValueCount = 0;
+    let criteriaCount = 0;
+
+    data.forEach((map) => {
+        const criteriaMap = map || new Map;
+        Object.keys(criteriaMap).forEach((criteriaKey) => {
+            const criteria = criteriaMap[criteriaKey] || {};
+            if (criteria.type === "label" || criteria.type === undefined) {
+                const values = criteria.values || [];
+                let num = 0;
+                Object.keys(values).forEach(valueKey => {
+                    const value = values[valueKey] || {};
+                    if (value.class === undefined && value.color === undefined && value.backgroundColor === undefined) {
+                        criteriaValueCount++;
+                        num++;
+                    }
+                });
+                if (num > 0) {
+                    criteriaSet.add(criteriaKey);
+                    criteriaCount++;
+                }
+            }
+        })
+    });
+
+    Object.keys(autoCriteria).forEach(criteriaKey => {
+        const criteria = autoCriteria[criteriaKey] || {};
+        if (criteria.type === "label" || criteria.type === undefined) {
+            const values = criteria.values || [];
+            let num = 0;
+            Object.keys(values).forEach(valueKey => {
+                const value = values[valueKey] || {};
+                if (value.class === undefined && value.color === undefined && value.backgroundColor === undefined) {
+                    criteriaValueCount++;
+                    num++;
+                }
+            });
+            if (num > 0 && !criteriaSet.has(criteriaKey)) {
+                criteriaCount++;
+            }
+        }
+    });
+
+    let delta = Math.floor(colorArray.length / criteriaValueCount);
+    let columnDelta;
+    if (delta < 1) {
+        columnDelta = Math.floor(colorArray.length / criteriaCount);
+    } else {
+        columnDelta = 0;
+    }
+
+    color = Math.floor(Math.random() * colorArray.length);
+
+    input.autoColor = input.autoColor || {};
+    let autoColor = input.autoColor;
+
+    data.forEach((map) => {
+        const criteriaMap = map || new Map;
+        completeAutoColor(criteriaMap);
+    });
+
+    completeAutoColor(autoCriteria);
+
+    function completeAutoColor(cmap) {
+        Object.keys(cmap).forEach((criteriaKey) => {
+            const criteria = cmap[criteriaKey] || {};
+            if (criteria.type === "label" || criteria.type === undefined) {
+                const values = criteria.values || [];
+                let tmpValues = [];
+                Object.keys(values).forEach((key) => {
+                    const value = values[key] || {};
+                    if (value.class === undefined && value.color === undefined && value.backgroundColor === undefined) {
+                        let obj = value;
+                        obj.name = key;
+                        tmpValues.push(obj);
+                    }
+                });
+
+                if (tmpValues[0] && tmpValues[0].weight === undefined) {
+                    tmpValues.sort((val1, val2) => {
+                        const name1 = val1.name || "";
+                        const name2 = val2.name || "";
+                        return name1.toString().localeCompare(name2.toString());
+                    })
+                } else {
+                    // TODO support order in config (ASC|DESC)
+                    let weight = 1;
+                    tmpValues.sort((val1, val2) => {
+                        const weight1 = val1.weight || 0;
+                        const weight2 = val2.weight || 0;
+                        return weight * (weight1 - weight2);
+                    });
+                }
+                if (tmpValues.length > 0) {
+                    autoColor[criteriaKey] = autoColor[criteriaKey] || {};
+                }
+
+                tmpValues.forEach(value => {
+                    if (!autoColor[criteriaKey][value.name]) {
+                        changed = true;
+                        autoColor[criteriaKey][value.name] = {
+                            color: foregroundArray[color],
+                            backgroundColor: colorArray[color]
+                        };
+                        color = (color + delta) % colorArray.length;
+                    }
+                });
+                color = (color + columnDelta) % colorArray.length;
+            }
+        });
+    }
+
+    if (changed) {
+        writeFileSync(config, yaml2json.safeDump(input), "utf8");
+    }
+    done();
+}
