+++ package.json
@@ -6,7 +6,7 @@
         "webpack:start": "webpack-dev-server --open --inline --progress --port 3000",
         "webpack:build": "webpack --config lib/webpack/webpack.common.js --progress --profile --bail",
         "gulp:compile": "node ./node_modules/gulp/bin/gulp.js default --gulpfile=./lib/gulp/gulpfile.js",
-        "gulp:development": "node ./node_modules/gulp/bin/gulp.js dev",
+        "gulp:development": "node ./node_modules/gulp/bin/gulp.js dev --gulpfile=./lib/gulp/gulpfile.js",
         "dev": "concurrently \"npm run gulp:development\" \"npm run webpack:start\" ",
         "build": "npm run gulp:compile && npm run webpack:build",
         "release": "node ./lib/gulp/publish.js"
@@ -74,7 +74,6 @@
         "rxjs": "^5.5.6",
         "showdown": "^1.8.6",
         "style": "0.0.3",
-        "sync-request": "^5.0.0",
         "tslint": "^5.9.1",
         "typescript": "2.6.2",
         "webpack": "^3.11.0",
