+++ src/app/components/comparison/configuration/configuration.service.ts
@@ -52,7 +52,7 @@ export class ConfigurationService {
 
 
     public loadComparison(cd: ChangeDetectorRef) {
-        this.http.get('comparison.yml', {responseType: 'text'})
+        this.http.get('comparison-auto-config.yml', {responseType: 'text'})
             .subscribe(res => {
                 const comparisonObject: any = yaml.safeLoad(res) || {};
                 const detailsObject: any = comparisonObject.details || {};
@@ -94,7 +94,7 @@ export class ConfigurationService {
                 );
 
                 /**
-                 * Construct map of criteria from 'comparison.yml'.criteria
+                 * Construct map of criteria from 'comparison-auto-config.yml'.criteria
                  */
                 const criteria: Map<string, Criteria> = new Map<string, Criteria>();
                 criteriaArray.forEach((obj) => Object.keys(obj).forEach((key) => {
@@ -159,7 +159,7 @@ export class ConfigurationService {
                 }));
 
                 /**
-                 * Complete map of criteria with 'comparison.yml'.autoCriteria
+                 * Complete map of criteria with 'comparison-auto-config.yml'.autoCriteria
                  */
                 Object.keys(autoCriteria).forEach((key) => {
                     const autoCriteriaObject = autoCriteria[key];
@@ -167,7 +167,7 @@ export class ConfigurationService {
                     const autoColorCriteria = isNullOrUndefined(autoColor[key]) ? {} : autoColor[key];
 
                     /**
-                     * If criteria is already defined by 'comparison.yml'.criteria
+                     * If criteria is already defined by 'comparison-auto-config.yml'.criteria
                      * complete criteria fields
                      */
                     if (criteria.get(key)) {
@@ -175,9 +175,9 @@ export class ConfigurationService {
                         let values: Map<string, CriteriaValue> = old.values;
                         /**
                          * Check each element CriteriaValue
-                         * oldValue from 'comparison.yml'.criteria
-                         * value from 'comparison.yml'.autoCriteria
-                         * color information from 'comparison.yml'.autoColorCriteria
+                         * oldValue from 'comparison-auto-config.yml'.criteria
+                         * value from 'comparison-auto-config.yml'.autoCriteria
+                         * color information from 'comparison-auto-config.yml'.autoColorCriteria
                          */
                         Object.keys(valuesObject).forEach(valueKey => {
                             const oldValue: CriteriaValue = old.values.get(valueKey);
@@ -226,7 +226,7 @@ export class ConfigurationService {
                             .setValues(values)
                             .build());
                         /**
-                         * If criteria is not defined by 'comparison.yml'.criteria use 'comparison.yml'.autoCriteria
+                         * If criteria is not defined by 'comparison-auto-config.yml'.criteria use 'comparison-auto-config.yml'.autoCriteria
                          */
                     } else {
                         let values: Map<string, CriteriaValue> = new Map<string, CriteriaValue>();
