+++ src/assets/.gitignore
@@ -1,4 +1,4 @@
 data.json
 description.md
-comparison.yml
+comparison-auto-config.yml
 style.css
